Module fix (#14)
* fix: test_example fixed and minor logic changes
* refactor: experiment call from console
* fix: correct directory generating
* refactor: experiments example moved to examples
